+++ client/components/Business/business-home.js
@@ -1,0 +1,34 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import {connect} from 'react-redux'
+
+/**
+ * COMPONENT
+ */
+export const BusinessHome = props => {
+  const {email} = props
+
+  return (
+    <div>
+      <h3>Welcome, {email}</h3>
+    </div>
+  )
+}
+
+/**
+ * CONTAINER
+ */
+const mapState = state => {
+  return {
+    email: state.user.email
+  }
+}
+
+export default connect(mapState)(BusinessHome)
+
+/**
+ * PROP TYPES
+ */
+BusinessHome.propTypes = {
+  email: PropTypes.string
+}
