+++ client/components/Customer/customer-home.js
@@ -5,7 +5,7 @@ import {connect} from 'react-redux'
 /**
  * COMPONENT
  */
-export const UserHome = props => {
+export const CustomerHome = props => {
   const {email} = props
 
   return (
@@ -24,11 +24,11 @@ const mapState = state => {
   }
 }
 
-export default connect(mapState)(UserHome)
+export default connect(mapState)(CustomerHome)
 
 /**
  * PROP TYPES
  */
-UserHome.propTypes = {
+CustomerHome.propTypes = {
   email: PropTypes.string
 }
