+++ server/db/models/index.js
@@ -11,7 +11,10 @@ Business.hasMany(Tier)
 Subscription.belongsTo(Tier)
 Tier.hasOne(Subscription)
 User.hasOne(Customer)
+Customer.belongsTo(User)
 User.hasOne(Business)
+Business.belongsTo(User)
+
 // register models
 // require('./models')
